refactor: reorganizar componentes de envio de OTP e atualizar chamadas de mutação
